Replace Detours with MinHook
This change removes the dependency on the proprietary Microsoft Detours
and replaces it with MinHook, which is BSD licensed.

This will allow making releases in accordance with the GPL.

This commit adds a binary release of MinHook. It is a release build by
VS2012, made from commit fea705c5e545c1ddd1c791d28c19db82121b9ec9 of
https://github.com/RaMMicHaeL/minhook.git .