Fix crash for DOF blur at default DOF resolution
Initialization of GAUSS (the DOF blur filter) would fail and cause a crash
at startup if dofBlurAmount was set to anything larger than zero and the
dofOverrideResolution was zero or unspecified.

This change treats the DOF resolution as 360 when configured as zero for
the purpose of initializing the blur filter and detecting the DOF render
targets.