Fix drawing of sun and lens flares
Whether the sun and lens flares are drawn are determined by the result of
an occlusion query. If the query returns a larger number of pixels drawn
than expected, which often happens when running in a higher resolution
than the default, the sun and lens flares might not be drawn.

This change wraps the IDirect3DQuery9 objects used for occlusion queries
and modifies the returned number of pixels drawn to correspond with the
default resolution, which makes the sun and lens flares appear as
expected.