Fix texture dumping overwriting .fx files
Texture loading, and thereby texture dumping is performed in a secondary
thread. Both texture dumping and .fx file loading makes calls to
GetDirectoryFile, which returns a pointer to a static buffer. Since these
activities can happen at the same time this is a race condition, and
sometimes .fx files are overwritten with dumped textures.

This change declares the static buffer for GetDirectoryFile as thread
local, which fixes the race condition.