Fix occlusion queries affected by antialiasing
Occlusion queries might return larger values than expected if
driver-enforced antialiasing is enabled.

This change makes an occlusion query for a square of a known size to
determine what result that should be expected. This is used by the
IDirect3DQuery9 wrapper to scale the result of further queries
appropriately, considering both the rendering resolution and unexpected
multisampling enforced by the driver.